address comments
* crosschainbank -> crosschain with aliases
* add command structure similar to other subcommands with bank >
  balances
* add spacing between print output
* default shows the denom: balance, can be overridden with `--combined`
  for a global balance view
