+++ main.go
@@ -16,12 +16,9 @@ limitations under the License.
 package main
 
 import (
-	"log"
-
 	"github.com/strangelove-ventures/lens/cmd"
 )
 
 func main() {
-	log.SetFlags(log.LstdFlags | log.Lshortfile)
 	cmd.Execute()
 }
